requirements-prod.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
poe-api-wrapper==1.7.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0